requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile
#
amqp==5.1.1
    # via kombu
asgiref==3.5.2
    # via django
async-timeout==4.0.2
    # via redis
billiard==3.6.4.0
    # via celery
celery[redis]==5.2.7
    # via
    #   -r requirements.in
    #   django-celery-beat
    #   django-celery-results
certifi==2022.9.24
    # via requests
charset-normalizer==2.1.1
    # via requests
click==8.1.3
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.3.0
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.2.0
    # via celery
crispy-bootstrap5==0.7
    # via -r requirements.in
deprecated==1.2.13
    # via redis
django==4.1.3
    # via
    #   -r requirements.in
    #   crispy-bootstrap5
    #   django-celery-beat
    #   django-filter
    #   djangorestframework
django-celery-beat==2.4.0
    # via -r requirements.in
django-celery-results==2.4.0
    # via -r requirements.in
django-crispy-forms==1.14.0
    # via
    #   -r requirements.in
    #   crispy-bootstrap5
django-filter==22.1
    # via -r requirements.in
django-timezone-field==5.0
    # via django-celery-beat
djangorestframework==3.14.0
    # via -r requirements.in
gunicorn==20.1.0
    # via -r requirements.in
idna==3.4
    # via requests
kombu==5.2.4
    # via celery
packaging==21.3
    # via redis
prompt-toolkit==3.0.32
    # via click-repl
psycopg2==2.9.5
    # via -r requirements.in
pyparsing==3.0.9
    # via packaging
python-crontab==2.6.0
    # via django-celery-beat
python-dateutil==2.8.2
    # via python-crontab
pytz==2022.6
    # via
    #   celery
    #   django-timezone-field
    #   djangorestframework
redis==4.3.4
    # via celery
requests==2.28.1
    # via -r requirements.in
six==1.16.0
    # via
    #   click-repl
    #   python-dateutil
sqlparse==0.4.3
    # via django
tzdata==2022.6
    # via django-celery-beat
urllib3==1.26.12
    # via requests
vine==5.0.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.5
    # via prompt-toolkit
whitenoise==6.2.0
    # via -r requirements.in
wrapt==1.14.1
    # via deprecated

# The following packages are considered to be unsafe in a requirements file:
# setuptools
